refactor: changes for the feedback
